backend/pyproject.toml
[project]
name = "fashion-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for a fashion pre-order MVP."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115,<1.0",
    "passlib[bcrypt]>=1.7,<2.0",
    "pydantic-settings>=2.7,<3.0",
    "python-jose[cryptography]>=3.3,<4.0",
    "python-multipart>=0.0.9,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn[standard]>=0.34,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "httpx>=0.28,<1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
